Show loading animation throught site
- Add pending on Legends, MultiSelectInput, SelectInput
- Show Legends while loading as well
- Make response persistent by default unless error/success
- Break Sankey and Table component for ProgramDetails
