+++ src/components/Backdrop/styles.css
@@ -5,9 +5,11 @@
     left: 0;
     align-items: center;
     justify-content: center;
+    z-index: 10 !important;
     background-color: var(--color-backdrop-light);
     width: 100%;
     height: 100%;
+    font-size: var(--font-size-extra-large);
 
     &.dark {
         background-color: var(--color-backdrop-dark);
