+++ src/components/BubbleLegend/styles.css
@@ -1,60 +1,65 @@
-.heading {
-    margin-bottom: var(--spacing-large-medium);
-}
+.bubble-legend {
+    position: relative;
+    min-height: 150px;
 
-.legend-item {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin: var(--spacing-small) 0;
-    height: var(--font-size-medium);
-    overflow: hidden;
-
-    .icon {
-        flex-shrink: 0;
-        margin-right: var(--spacing-small);
-        font-size: var(--font-size-large);
+    .heading {
+        margin-bottom: var(--spacing-large-medium);
     }
 
-    .color {
-        flex-shrink: 0;
-        opacity: 0.6;
-        margin-right: var(--spacing-small);
-        border-radius: 50%;
-        width: var(--font-size-medium);
-        height: var(--font-size-medium);
-    }
-
-    .value {
-        flex-grow: 1;
-        margin-left: var(--spacing-small);
-        text-transform: capitalize;
-        color: var(--color-text-label);
-    }
-}
-
-.footer {
-    display: flex;
-    flex-direction: column;
-
-    .circle {
-        opacity: 0.6;
-        margin-right: var(--spacing-small);
-        border-radius: 50%;
-        width: var(--font-size-medium);
-        height: var(--font-size-medium);
-        content: '';
-    }
-
-    .positive,
-    .negative {
+    .legend-item {
         display: flex;
         align-items: center;
-        flex-shrink: 0;
-        font-size: var(--font-size-small);
+        justify-content: center;
+        margin: var(--spacing-small) 0;
+        height: var(--font-size-medium);
+        overflow: hidden;
+
+        .icon {
+            flex-shrink: 0;
+            margin-right: var(--spacing-small);
+            font-size: var(--font-size-large);
+        }
+
+        .color {
+            flex-shrink: 0;
+            opacity: 0.6;
+            margin-right: var(--spacing-small);
+            border-radius: 50%;
+            width: var(--font-size-medium);
+            height: var(--font-size-medium);
+        }
+
+        .value {
+            flex-grow: 1;
+            margin-left: var(--spacing-small);
+            text-transform: capitalize;
+            color: var(--color-text-label);
+        }
     }
 
-    .negative {
-        margin-bottom: var(--spacing-small);
+    .footer {
+        display: flex;
+        flex-direction: column;
+
+        .circle {
+            opacity: 0.6;
+            margin-right: var(--spacing-small);
+            border-radius: 50%;
+            width: var(--font-size-medium);
+            height: var(--font-size-medium);
+            content: '';
+        }
+
+        .positive,
+        .negative {
+            display: flex;
+            align-items: center;
+            flex-shrink: 0;
+            font-size: var(--font-size-small);
+        }
+
+        .negative {
+            margin-bottom: var(--spacing-small);
+        }
     }
 }
