+++ src/components/ChoroplethLegend/styles.css
@@ -1,4 +1,7 @@
 .legend-container {
+    position: relative;
+    min-height: 100px;
+
     .heading {
         margin-bottom: 0;
     }
