+++ src/views/Dashboard/ProgramDetails/Sankey/styles.css
@@ -1,0 +1,6 @@
+.sankey {
+    position: relative;
+    flex-grow: 1;
+    margin: var(--spacing-small) var(--spacing-medium);
+    /* overflow-y: auto; */
+}
