+++ src/views/Dashboard/index.tsx
@@ -303,21 +303,25 @@ const Dashboard = (props: Props) => {
     const [
         indicatorMapStatePending,
         indicatorMapState,
-    ] = useMapStateForIndicator(regionLevel, validSelectedIndicator);
+    ] = useMapStateForIndicator(regionLevel, validSelectedIndicator, false);
 
     const [
         fiveWMapStatePending,
         fiveWMapState,
         fiveWStats,
-    ] = useMapStateForFiveW(regionLevel, programs, selectedFiveWOption);
+    ] = useMapStateForFiveW(regionLevel, programs, selectedFiveWOption, false);
 
     const {
+        choroplethSelected,
         choroplethMapState,
+        choroplethPending,
         choroplethTitle,
         choroplethInteger,
         choroplethUnit,
 
+        bubbleSelected,
         bubbleMapState,
+        bubblePending,
         bubbleTitle,
         bubbleInteger,
         bubbleUnit,
@@ -332,25 +336,33 @@ const Dashboard = (props: Props) => {
         if (mapStyleInverted) {
             return {
                 choroplethMapState: indicatorMapState,
+                choroplethPending: indicatorMapStatePending,
                 choroplethTitle: selectedIndicatorDetails?.fullTitle,
                 choroplethInteger: selectedIndicatorDetails?.datatype === 'integer',
                 choroplethUnit: selectedIndicatorDetails?.unit,
+                choroplethSelected: isDefined(selectedIndicator),
 
                 bubbleMapState: fiveWMapState,
+                bubblePending: fiveWMapStatePending,
                 bubbleTitle: fiveWTitle,
                 bubbleInteger: fiveW?.datatype === 'integer',
                 bubbleUnit: fiveW?.unit,
+                bubbleSelected: isDefined(selectedFiveWOption),
 
                 titleForPrintBar: title,
             };
         }
         return {
+            choroplethSelected: isDefined(selectedFiveWOption),
             choroplethMapState: fiveWMapState,
+            choroplethPending: fiveWMapStatePending,
             choroplethTitle: fiveWTitle,
             choroplethInteger: fiveW?.datatype === 'integer',
             choroplethUnit: fiveW?.unit,
 
+            bubbleSelected: isDefined(selectedIndicator),
             bubbleMapState: indicatorMapState,
+            bubblePending: indicatorMapStatePending,
             bubbleTitle: selectedIndicatorDetails?.fullTitle,
             bubbleInteger: selectedIndicatorDetails?.datatype === 'integer',
             bubbleUnit: selectedIndicatorDetails?.unit,
@@ -361,7 +373,10 @@ const Dashboard = (props: Props) => {
         mapStyleInverted,
         indicatorMapState,
         fiveWMapState,
+        indicatorMapStatePending,
+        fiveWMapStatePending,
         selectedFiveWOption,
+        selectedIndicator,
         selectedIndicatorDetails,
     ]);
 
@@ -435,6 +450,7 @@ const Dashboard = (props: Props) => {
 
     const hash = useHash();
 
+    /*
     const showLegend = (
         bubbleLegend.length > 0
         || Object.keys(mapLegend).length > 0
@@ -442,6 +458,7 @@ const Dashboard = (props: Props) => {
         || selectedRasterLayerDetail
         || (selectedVectorLayersDetail && selectedVectorLayersDetail?.length > 0)
     );
+    */
 
     const dfidData = useMemo(
         () => {
@@ -733,6 +750,7 @@ const Dashboard = (props: Props) => {
                     optionLabelSelector={indicatorLabelSelector}
                     optionKeySelector={indicatorKeySelector}
                     groupKeySelector={indicatorGroupKeySelector}
+                    pending={indicatorListPending}
                 />
                 {selectedIndicatorDetails && selectedIndicatorDetails.abstract && (
                     <div className={styles.abstract}>
@@ -767,8 +785,8 @@ const Dashboard = (props: Props) => {
                     optionLabelSelector={layerLabelSelector}
                 />
             </div>
-            {showLegend && (
-                <div className={styles.legendContainer}>
+            <div className={styles.legendContainer}>
+                {choroplethSelected && (
                     <ChoroplethLegend
                         className={styles.legend}
                         title={choroplethTitle}
@@ -777,7 +795,10 @@ const Dashboard = (props: Props) => {
                         unit={choroplethUnit}
                         minExceeds={dataMinExceeds}
                         maxExceeds={dataMaxExceeds}
+                        pending={choroplethPending}
                     />
+                )}
+                {bubbleSelected && (
                     <BubbleLegend
                         className={styles.legend}
                         title={bubbleTitle}
@@ -787,45 +808,46 @@ const Dashboard = (props: Props) => {
                         radiusSelector={legendRadiusSelector}
                         legendType={bubbleLegendType}
                         unit={bubbleUnit}
+                        pending={bubblePending}
                     />
-                    {selectedRasterLayerDetail && (
-                        <RasterLegend
-                            className={styles.legend}
-                            rasterLayer={selectedRasterLayerDetail}
-                        />
-                    )}
-                    {selectedVectorLayersDetail && selectedVectorLayersDetail.length > 0 && (
-                        <VectorLegend
-                            className={styles.legend}
-                            vectorLayers={selectedVectorLayersDetail}
-                        />
-                    )}
-                    {showTravelTimeChoropleth && (
-                        <>
-                            {selectedTravelTimeType === 'catchment' && (
-                                <ChoroplethLegend
-                                    title="Catchment"
-                                    className={styles.legend}
-                                    minValue=""
-                                    opacity={0.6}
-                                    unit="hours"
-                                    legend={catchmentLegend}
-                                />
-                            )}
-                            {selectedTravelTimeType === 'uncovered' && (
-                                <ChoroplethLegend
-                                    title="Uncovered"
-                                    className={styles.legend}
-                                    minValue=""
-                                    opacity={0.6}
-                                    unit="hours"
-                                    legend={uncoveredLegend}
-                                />
-                            )}
-                        </>
-                    )}
-                </div>
-            )}
+                )}
+                {selectedRasterLayerDetail && (
+                    <RasterLegend
+                        className={styles.legend}
+                        rasterLayer={selectedRasterLayerDetail}
+                    />
+                )}
+                {selectedVectorLayersDetail && selectedVectorLayersDetail.length > 0 && (
+                    <VectorLegend
+                        className={styles.legend}
+                        vectorLayers={selectedVectorLayersDetail}
+                    />
+                )}
+                {showTravelTimeChoropleth && (
+                    <>
+                        {selectedTravelTimeType === 'catchment' && (
+                            <ChoroplethLegend
+                                title="Catchment"
+                                className={styles.legend}
+                                minValue=""
+                                opacity={0.6}
+                                unit="hours"
+                                legend={catchmentLegend}
+                            />
+                        )}
+                        {selectedTravelTimeType === 'uncovered' && (
+                            <ChoroplethLegend
+                                title="Uncovered"
+                                className={styles.legend}
+                                minValue=""
+                                opacity={0.6}
+                                unit="hours"
+                                legend={uncoveredLegend}
+                            />
+                        )}
+                    </>
+                )}
+            </div>
             <PrintDetailsBar
                 show={printMode}
                 title={titleForPrintBar}
@@ -833,6 +855,7 @@ const Dashboard = (props: Props) => {
             />
             {hash === 'regions' && (
                 <RegionDetails
+                    indicatorListPending={indicatorListPending}
                     indicatorList={indicatorList}
                 />
             )}
